+++ FE/src/app/components/organisation/organisation.component.ts
@@ -11,13 +11,14 @@ import { EditOrganisationDialog } from './edit-organisation/edit-organisation-di
 import { RemoveOrganisationDialog } from './remove-organisation/remove-organisation-dialog';
 import { Item } from 'src/app/models/Item';
 import { FirebaseService } from 'src/app/services/firebase/firebase.service';
-import { ItemService } from 'src/app/services/firebase/item.service'
+import { ItemService } from 'src/app/services/firebase/item.service';
 import { Organisation } from 'src/app/models/Organisation/Organisation';
 import { Subscription } from 'rxjs';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { AddDonationItemComponent } from '../donation-item/add-donation-item/add-donation-item.component';
 import { RemoveDonationItemComponent } from '../donation-item/remove-donation-item/remove-donation-item.component';
 import { UpdateItemsComponent } from '../donation-item/update-donation-item/update-donation-item.component';
+import { OrganisationService } from 'src/app/services/firebase/organisation-service/organisation.service';
 
 @Component({
   selector: 'app-organisation',
@@ -62,8 +63,6 @@ export class OrganisationComponent {
   //snackbar variables
   message: string;
 
-
-
   @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
   @ViewChild(MatSort) sort: MatSort;
 
@@ -73,26 +72,25 @@ export class OrganisationComponent {
     public http: HttpClient,
     public storage: AngularFireStorage,
     public fs: FirebaseService,
+    public ofs: OrganisationService,
     public _snackBar: MatSnackBar,
     public ifs: ItemService
-
-  ) { }
+  ) {}
 
   ngOnDestroy(): void {
     this.getOrgsSubscription.unsubscribe();
     this.getItemsSubscription.unsubscribe();
-
   }
 
   ngOnInit(): void {
     //this.ifs.addItem('4H9O58oiEH0D88AXZEUF',{name:"hi",activeStatus:false,description:"hi",summary:"hi",initialPrice:20,createdAt:serverTimestamp(), img:"",totalDonations:0, dateCompleted:null})
     this.getOrgs();
     this.initSelectedOrg();
-
   }
 
   onImgError(event) {
-    event.target.src = 'https://freepikpsd.com/file/2019/10/placeholder-image-png-5-Transparent-Images.png'
+    event.target.src =
+      'https://freepikpsd.com/file/2019/10/placeholder-image-png-5-Transparent-Images.png';
   }
 
   initSelectedOrg() {
@@ -108,10 +106,9 @@ export class OrganisationComponent {
       summary: '',
       totalDonationItems: 0,
       totalDonations: 0,
-      website: ''
-    }
+      website: '',
+    };
   }
-
 
   //add form validation
   addDonationItemDialog(): void {
@@ -119,15 +116,13 @@ export class OrganisationComponent {
       width: '730px',
       data: {
         id: this.selectedOrg.id,
-        file: this.file
+        file: this.file,
       },
     });
 
     dialogRef.afterClosed().subscribe(async (result: any) => {
-
       if (result.file) {
-
-        this.fs.uploadImage(this.selectedOrg.id, result.file, result.itemRef)
+        this.fs.uploadImage(this.selectedOrg.id, result.file, result.itemRef);
 
         //   this.fs.addDonationItem(result).then ((response) => {
         //     this.openSnackBar(response.message)
@@ -142,23 +137,23 @@ export class OrganisationComponent {
       data: {
         itemID: itemID,
         id: this.selectedOrg.id,
-        itemName: itemName
+        itemName: itemName,
       },
     });
 
     dialogRef.afterClosed().subscribe(async (result: any) => {
       //----------------------------- Remove a Donation Item --------------------------//
       if (result.isDeleted === true) {
+        this.storage
+          .ref(
+            `Organisations/${this.selectedOrg.id}/Items/${result.itemID}/itemImg`
+          )
+          .delete();
 
-        this.storage.ref(`Organisations/${this.selectedOrg.id}/Items/${result.itemID}/itemImg`).delete()
-
-        this.openSnackBar(itemName + " successfully deleted")
+        this.openSnackBar(itemName + ' successfully deleted');
       }
-    })
-
-
+    });
   }
-
 
   addOrgDialog(): void {
     const dialogRef = this.dialog.open(AddOrganisationDialog, {
@@ -181,9 +176,9 @@ export class OrganisationComponent {
     dialogRef.afterClosed().subscribe(async (result: any) => {
       //----------------------------- Create an Org --------------------------//
       if (result) {
-        this.fs.addOrganisation(result).then((response) => {
-          this.openSnackBar(response.message)
-        })
+        this.ofs.addOrganisation(result).then((response) => {
+          this.openSnackBar(response.message);
+        });
       }
     });
   }
@@ -210,31 +205,32 @@ export class OrganisationComponent {
     // runs after dialog closes. updates org
     dialogRef.afterClosed().subscribe((result) => {
       if (result) {
-
         const orgReq: Organisation = {
           id: result.id,
           ABN: result.ABN,
-          description: result.description ? result.description : "",
-          name: result.name ? result.name : "",
-          phone: result.phone ? result.phone : "",
-          summary: result.summary ? result.summary : "",
-          website: result.website ? result.website : "",
-          img: result.img ? result.img : "",
-          totalDonationItems: result.totalDonationItems ? result.totalDonationItems : 0,
+          description: result.description ? result.description : '',
+          name: result.name ? result.name : '',
+          phone: result.phone ? result.phone : '',
+          summary: result.summary ? result.summary : '',
+          website: result.website ? result.website : '',
+          img: result.img ? result.img : '',
+          totalDonationItems: result.totalDonationItems
+            ? result.totalDonationItems
+            : 0,
           totalDonations: result.totalDonations ? result.totalDonations : 0,
-          activeStatus: result.activeStatus
-        }
+          activeStatus: result.activeStatus,
+        };
 
-        this.fs.editOrganisation(this.selectedOrg.id, orgReq)
-          .then((resp) => {
-            this.selectedOrg = resp
-            this.openSnackBar(resp.name + " Edited Successfully")
+        this.ofs.editOrganisation(this.selectedOrg.id, orgReq).then((resp) => {
+          this.selectedOrg = resp;
+          this.openSnackBar(resp.name + ' Edited Successfully');
 
-            if (result?.file) {
-              this.fs.uploadImage(this.selectedOrg.id, result.file)
-                .then((imgURL) => this.selectedOrg.img = imgURL)
-            }
-          })
+          if (result?.file) {
+            this.fs
+              .uploadImage(this.selectedOrg.id, result.file)
+              .then((imgURL) => (this.selectedOrg.img = imgURL));
+          }
+        });
       }
     });
   }
@@ -246,17 +242,15 @@ export class OrganisationComponent {
         id: this.selectedOrg.id,
         name: this.selectedOrg.name,
         totalDonationItems: this.selectedOrg.totalDonationItems,
-        totalDonations: this.selectedOrg.totalDonations
+        totalDonations: this.selectedOrg.totalDonations,
       },
     });
 
     dialogRef.afterClosed().subscribe((result) => {
-
       if (result === true) {
-
-        this.fs.removeOrganisation(this.selectedOrg.id).then((response) => {
-          this.openSnackBar(response.message)
-        })
+        this.ofs.removeOrganisation(this.selectedOrg.id).then((response) => {
+          this.openSnackBar(response.message);
+        });
       }
     });
   }
@@ -264,21 +258,32 @@ export class OrganisationComponent {
   toggleActiveStatus() {
     this.initSelectedOrg();
     this.activeStatusToggle = !this.activeStatusToggle;
-    this.getOrgsSubscription = this.fs.getOrgs(this.activeStatusToggle)
-      .subscribe(
-        orgs => {
-          this.orgData = new MatTableDataSource(orgs);
-          this.orgData.paginator = this.paginator;
-          this.orgData.sort = this.sort;
-          this.orgData.filterPredicate = function (data, filter: string): boolean {
-            return data.name.trim().toLowerCase().includes(filter) ||
-              data.totalDonations.toString().trim().toLowerCase().includes(filter) ||
-              data.totalDonationItems.toString().trim().toLowerCase().includes(filter);
-          };
-        })
-
+    this.getOrgsSubscription = this.ofs
+      .getOrgs(this.activeStatusToggle)
+      .subscribe((orgs) => {
+        this.orgData = new MatTableDataSource(orgs);
+        this.orgData.paginator = this.paginator;
+        this.orgData.sort = this.sort;
+        this.orgData.filterPredicate = function (
+          data,
+          filter: string
+        ): boolean {
+          return (
+            data.name.trim().toLowerCase().includes(filter) ||
+            data.totalDonations
+              .toString()
+              .trim()
+              .toLowerCase()
+              .includes(filter) ||
+            data.totalDonationItems
+              .toString()
+              .trim()
+              .toLowerCase()
+              .includes(filter)
+          );
+        };
+      });
   }
-
 
   // Function to update item called in the dialog component
   openItemUpdateDialog(item: Item): void {
@@ -288,38 +293,51 @@ export class OrganisationComponent {
       height: 'fit-content',
       maxHeight: '90vh',
       data: {
-        item: item, org: this.selectedOrg.id
-      }
+        item: item,
+        org: this.selectedOrg.id,
+      },
     });
 
     dialogRef.afterClosed().subscribe((res) => {
       if (res?.file) {
-        this.fs.uploadImage(this.selectedOrg.id, res.file, res.item.id)
+        this.fs.uploadImage(this.selectedOrg.id, res.file, res.item.id);
       }
-    })
+    });
   }
 
   getOrgs() {
-
-    this.getOrgsSubscription = this.fs.getOrgs(this.activeStatusToggle)
-      .subscribe(orgs => {
+    this.getOrgsSubscription = this.ofs
+      .getOrgs(this.activeStatusToggle)
+      .subscribe((orgs) => {
         this.orgData = new MatTableDataSource(orgs);
         this.orgData.paginator = this.paginator;
         this.orgData.sort = this.sort;
-        this.orgData.filterPredicate = function (data, filter: string): boolean {
-          return data.name.trim().toLowerCase().includes(filter) ||
-            data.totalDonations.toString().trim().toLowerCase().includes(filter) ||
-            data.totalDonationItems.toString().trim().toLowerCase().includes(filter);
+        this.orgData.filterPredicate = function (
+          data,
+          filter: string
+        ): boolean {
+          return (
+            data.name.trim().toLowerCase().includes(filter) ||
+            data.totalDonations
+              .toString()
+              .trim()
+              .toLowerCase()
+              .includes(filter) ||
+            data.totalDonationItems
+              .toString()
+              .trim()
+              .toLowerCase()
+              .includes(filter)
+          );
         };
-      })
+      });
   }
 
   //-------------------- GET ITEMS --------------------\\
   getItems(orgID) {
-    this.getItemsSubscription = this.ifs.getItems(orgID)
-      .subscribe(items => {
-        this.items = items as Item[]
-      })
+    this.getItemsSubscription = this.ifs.getItems(orgID).subscribe((items) => {
+      this.items = items as Item[];
+    });
   }
 
   //-------------------- GET ITEMS --------------------\\
@@ -335,10 +353,9 @@ export class OrganisationComponent {
 
   // selected row of org table
   selectRow(orgData) {
-
     if (this.selectedOrg.id === orgData.id) {
       this.initSelectedOrg();
-      this.getItemsSubscription.unsubscribe()
+      this.getItemsSubscription.unsubscribe();
 
       return;
     }
