+++ FE/src/app/services/firebase/organisation-service/organisation.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { OrganisationService } from './organisation.service';
+
+describe('OrganisationService', () => {
+  let service: OrganisationService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(OrganisationService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
